requirements.txt
# requirements.txt
requests
streamlit
python-dotenv